requirements.txt
python-dotenv~=1.0.0
pinecone-client
langchain==0.0.335
langchain-experimental==0.0.40
deep-translator
PyMuPDF
docx2txt
openai~=1.3.0
tiktoken==0.3.3
django==4.2.7